refactor typography
