--- src/constants/fonts/index.ts
@@ -1,7 +1,0 @@
-export default {
-  headerFont:
-    "'nocturno', Georgia, TimesNewRoman, 'Times New Roman', Times, Baskerville, serif",
-  baseFont:
-    "'larssiet', 'Avant Garde', Avantgarde, 'Century Gothic', CenturyGothic, AppleGothic, Verdana, sans-serif",
-  monoFont: "'apercu mono', 'Lucida Console', Monaco, monospace",
-};
