+++ src/constants/themes/primaryTheme.ts
@@ -1,6 +1,6 @@
 import PRIMARY_COLORS from '../colors/primary';
 import PRIMARY_FONTS from '../fonts/primary';
-import { PRIMARY_TYPOGRAPHY } from '../typography/primaryTypography';
+import { PRIMARY_TYPOGRAPHY } from '../typography/primary';
 
 export const primaryTheme = {
   __type: 'primary',
