+++ src/constants/themes/secondaryTheme.ts
@@ -1,6 +1,6 @@
 import SECONDARY_COLORS from '../colors/secondary';
 import SECONDARY_FONTS from '../fonts/secondary';
-import { SECONDARY_TYPOGRAPHY } from '../typography/secondaryTypography';
+import { SECONDARY_TYPOGRAPHY } from '../typography/secondary';
 
 export const secondaryTheme = {
   __type: 'secondary',
