+++ src/shared-components/alert/style.ts
@@ -7,7 +7,6 @@ import {
   SPACER,
   ANIMATION,
   ThemeType,
-  TYPOGRAPHY_CONSTANTS,
 } from '../../constants';
 
 import { AlertType } from '.';
@@ -103,7 +102,7 @@ export const MainContainer = styled.div`
   justify-content: flex-start;
   align-items: flex-start;
   color: ${({ theme }) => theme.COLORS.white};
-  font-size: ${TYPOGRAPHY_CONSTANTS.fontSize.caption};
+  font-size: ${({ theme }) => theme.TYPOGRAPHY.fontSize.caption};
   padding: ${SPACER.medium};
 `;
 
@@ -120,7 +119,7 @@ export const CtaContent = styled.div`
   justify-content: space-between;
   align-items: center;
   color: ${({ theme }) => theme.COLORS.white};
-  font-size: ${TYPOGRAPHY_CONSTANTS.fontSize.caption};
+  font-size: ${({ theme }) => theme.TYPOGRAPHY.fontSize.caption};
   border-top: 1px solid rgba(255, 255, 255, 0.1);
 `;
