+++ src/shared-components/banner/style.ts
@@ -1,13 +1,7 @@
 import styled from '@emotion/styled';
 import { buttonReset } from 'src/utils/styles/buttonReset';
 
-import {
-  BOX_SHADOWS,
-  MEDIA_QUERIES,
-  SPACER,
-  ThemeType,
-  TYPOGRAPHY_CONSTANTS,
-} from '../../constants';
+import { BOX_SHADOWS, MEDIA_QUERIES, SPACER, ThemeType } from '../../constants';
 
 import { BannerType } from '.';
 
@@ -68,7 +62,7 @@ export const MainContainer = styled.div`
   justify-content: flex-start;
   align-items: flex-start;
   color: ${({ theme }) => theme.COLORS.white};
-  font-size: ${TYPOGRAPHY_CONSTANTS.fontSize.caption};
+  font-size: ${({ theme }) => theme.TYPOGRAPHY.fontSize.caption};
 `;
 
 export const ContentContainer = styled.div`
