+++ src/shared-components/dropdown/style.ts
@@ -1,13 +1,7 @@
 import styled from '@emotion/styled';
 import { css } from '@emotion/core';
 
-import {
-  BOX_SHADOWS,
-  SPACER,
-  ANIMATION,
-  ThemeType,
-  TYPOGRAPHY_CONSTANTS,
-} from '../../constants';
+import { BOX_SHADOWS, SPACER, ANIMATION, ThemeType } from '../../constants';
 
 type DropdownInputStyleProps = {
   borderRadius: string;
@@ -177,10 +171,10 @@ export const DropdownOption = styled.li<{
     box-shadow: ${BOX_SHADOWS.focusInner};
   }
 
-  ${({ selected }) =>
+  ${({ selected, theme }) =>
     selected &&
     `
-      font-weight: ${TYPOGRAPHY_CONSTANTS.fontWeight.bold};
+      font-weight: ${theme.TYPOGRAPHY.fontWeight.bold};
     `}
 
   ${({ disabled, theme }) =>
